Simulation of Memory Paging System
This is a simulation of memory paging system. When the memory/frame is empty and a page is loading in, a 'miss' counter goes up. If the same data requested is found in memory/frame, a 'hit' counter goes up. If the memory/frame is full and the page requested is not in memory/frame, a 'miss' counter goes up, the program will swap out a page in memory/frame with the data requested. A page-table is made to keep track of each page if it is present in memory/frame. The program will keep track of the number of times a page has been 'hit' and 'miss', depending on the size of the memory/frame the user has given it.